backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
sentence-transformers
scikit-learn
requests
python-dotenv
google-generativeai